round numbers for py3 tests

+++ colorlover/__init__.py
@@ -1643,7 +1643,7 @@ def scale_type( scale ):
 Colorscales must be in one of these 3 forms:\n\
 [ (255, 255, 255), (255, 255, 255), (255, 255, 255) ]\n\
 [ "rgb(255, 255, 255)", "rgb(255, 255, 255)", "rgb(255, 255, 255)" ]\n\
-[ "hsl(360,100,100)", "hsl(360,100,100)", "hsl(360,100,100)" ]') 
+[ "hsl(360,100,100)", "hsl(360,100,100)", "hsl(360,100,100)" ]')
 
 def to_numeric( scale ):
     ''' converts scale of rgb or hsl strings to list of tuples with rgb integer values. ie,
@@ -1660,22 +1660,22 @@ def to_numeric( scale ):
     return numeric_scale
 
 def to_hsl( scale ):
-    ''' convert a string rgb or numeric rgb colorscale to hsl. ie, 
+    ''' convert a string rgb or numeric rgb colorscale to hsl. ie,
         [ "rgb(255, 255, 255)", "rgb(255, 255, 255)", "rgb(255, 255, 255)" ] -->
         [ "hsl(360,100%,100%)", "hsl(360,100%,100%)", "hsl(360,100%,100%)" ]
         add percentages to saturation and lightness if missing for css compatibility '''
-    
+
     hsl = []
     s_t = scale_type(scale)
 
     if s_t == 'hsl':
         # add percentages to s and l if missing
-        
+
         numeric_hsl_scale = []
         for s in scale:
             s = s[s.find("(")+1:s.find(")")].replace(' ','').replace('%','').split(',')
             numeric_hsl_scale.append( ( float(s[0]), float(s[1]), float(s[2]) ) )
-        
+
         for ea in numeric_hsl_scale:
             h,s,l = [ str(x) for x in ea ]
             if s[-1] != '%':
@@ -1684,7 +1684,7 @@ def to_hsl( scale ):
                 l = l + '%'
             hsl_str = 'hsl(' + ', '.join([h,s,l]) + ')'
             hsl.append( hsl_str )
-                
+
         return hsl
 
     elif s_t == 'rgb':
@@ -1693,18 +1693,18 @@ def to_hsl( scale ):
     for ea in scale:
         r,g,b = [ x/255.0 for x in ea ]
         h,l,s = colorsys.rgb_to_hls( r,g,b )
-        h,s,l = [ str(round(h*360.0)), str(round(s*100.0))+'%', str(round(l*100.0))+'%' ]
+        h,s,l = [ str(int(round(h*360.0))), str(int(round(s*100.0)))+'%', str(int(round(l*100.0)))+'%' ]
         hsl_str = 'hsl(' + ', '.join([h,s,l]) + ')'
         hsl.append( hsl_str )
-                
+
     return hsl
 
 def to_rgb( scale ):
-    ''' convert an hsl or numeric rgb color scale to string rgb color scale. ie, 
+    ''' convert an hsl or numeric rgb color scale to string rgb color scale. ie,
         [ "hsl(360,100,100)", "hsl(360,100,100)", "hsl(360,100,100)" ] -->
         [ "rgb(255, 255, 255)", "rgb(255, 255, 255)", "rgb(255, 255, 255)" ]
         '''
-    
+
     rgb = []
     s_t = scale_type(scale)
 
@@ -1727,12 +1727,12 @@ def to_rgb( scale ):
         r,g,b = [ str(int(round(x*255.0))) for x in (r,g,b) ]
         rgb_str = 'rgb(' + ', '.join([r,g,b]) + ')'
         rgb.append( rgb_str )
-        
+
     return rgb
 
 def to_html( scale ):
-    ''' traverse color scale dictionary and return available color scales in HTML string '''  
-    
+    ''' traverse color scale dictionary and return available color scales in HTML string '''
+
     global s
     s = ''
 
@@ -1742,15 +1742,15 @@ def to_html( scale ):
             scale = to_rgb( scale )
         s_s = ''
         for ea in scale:
-             s_s+='<div style="background-color:{0};height:20px;width:20px;display:inline-block;"></div>'.format(ea)    
+             s_s+='<div style="background-color:{0};height:20px;width:20px;display:inline-block;"></div>'.format(ea)
         return s_s
-    
+
     def section_titles( k ):
         d = { 'qual':'Qualitative','div':'Diverging','seq':'Sequential' }
         if k in list(d.keys()):
             return '<h4>' + d[k] + '</h4>'
         return '<hr><h3>' + k + ' colors</h3>'
-    
+
     def prettyprint( d ):
         global s
         for k, v in list(d.items()):
@@ -1761,12 +1761,12 @@ def to_html( scale ):
             else:
                 s += '<div style="display:inline-block;padding:10px;"><div>{0}</div>{1}</div>'.format(k, single_scale( v ) )
         return s
-    
+
     if isinstance( scale, list ):
         return single_scale( scale )
     elif isinstance( scale, dict ):
         prettyprint( scale )
-    
+
     return s
 
 def flipper( scl=None ):
@@ -1779,18 +1779,18 @@ def flipper( scl=None ):
     return flipped
 
 def interp(scl, r):
-    ''' Interpolate a color scale "scl" to a new one with length "r" 
-        Fun usage in IPython notebook: 
+    ''' Interpolate a color scale "scl" to a new one with length "r"
+        Fun usage in IPython notebook:
         HTML( to_html( to_hsl( interp( cl.scales['11']['qual']['Paired'], 5000 ) ) ) ) '''
     c = []
-    SCL_FI = len(scl)-1 # final index of color scale 
+    SCL_FI = len(scl)-1 # final index of color scale
     r = [x * 0.1 for x in range(r)] if isinstance( r, int ) else r
     scl = to_numeric( scl )
 
     def interp3(fraction, start, end):
         ''' Interpolate between values of 2, 3-member tuples '''
         def intp(f, s, e):
-            return s + (e - s)*f 
+            return s + (e - s)*f
         return tuple([intp(fraction, start[i], end[i]) for i in range(3)])
 
     def rgb_to_hsl(rgb):
@@ -1806,7 +1806,7 @@ def interp(scl, r):
             h = 0
             s = 0 if l > 0 and l < 1 else h
         else:
-            d = mx - mn;        
+            d = mx - mn;
             s =  d / (mx + mn) if l < 0.5 else d / (2 - mx - mn)
             if mx == r:
                 h = (g - b) / d + ( 6 if g < b else 0 )
@@ -1815,19 +1815,19 @@ def interp(scl, r):
             else:
                 h = r - g / d + 4
 
-        return (round(h*60,4), round(s*100,4), round(l*100,4))
-    
+        return (int(round(h*60,4)), int(round(s*100,4)), int(round(l*100,4)))
+
     for i in r:
         c_i = int(i*math.floor(SCL_FI)/round(r[-1])) # start color index
         hsl_o = rgb_to_hsl( scl[c_i] ) # convert rgb to hls
-        hsl_f = rgb_to_hsl( scl[c_i+1] ) 
+        hsl_f = rgb_to_hsl( scl[c_i+1] )
         section_min = c_i*r[-1]/SCL_FI
         section_max = (c_i+1)*(r[-1]/SCL_FI)
         fraction = (i-section_min)/(section_max-section_min)
         hsl = interp3( fraction, hsl_o, hsl_f )
         c.append( 'hsl'+str(hsl) )
-        
+
     return to_hsl( c )
 
 
- 
+
